TEST: move `ThrowingReporter` and make sure bug is fixed
