+++ base/src/test/java/org/aya/concrete/DesugarTest.java
@@ -7,9 +7,9 @@ import org.aya.concrete.desugar.AyaBinOpSet;
 import org.aya.pretty.doc.Doc;
 import org.aya.resolve.ResolveInfo;
 import org.aya.resolve.context.EmptyContext;
-import org.aya.test.ThrowingReporter;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.error.Global;
+import org.aya.util.reporter.ThrowingReporter;
 import org.intellij.lang.annotations.Language;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
