+++ base/src/test/java/org/aya/concrete/ParseTest.java
@@ -11,11 +11,11 @@ import org.aya.concrete.stmt.Command;
 import org.aya.concrete.stmt.Decl;
 import org.aya.concrete.stmt.Stmt;
 import org.aya.pretty.doc.Doc;
-import org.aya.test.ThrowingReporter;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.error.Global;
 import org.aya.util.error.SourceFile;
 import org.aya.util.error.SourcePos;
+import org.aya.util.reporter.ThrowingReporter;
 import org.intellij.lang.annotations.Language;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
