+++ base/src/test/java/org/aya/test/LibraryTest.java
@@ -1,9 +1,10 @@
-// Copyright (c) 2020-2021 Yinsen (Tesla) Zhang.
+// Copyright (c) 2020-2022 Yinsen (Tesla) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
 package org.aya.test;
 
 import org.aya.cli.library.LibraryCompiler;
 import org.aya.util.FileUtil;
+import org.aya.util.reporter.ThrowingReporter;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
