+++ base/src/test/java/org/aya/test/StreamReporter.java
@@ -2,10 +2,10 @@
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
 package org.aya.test;
 
-import org.aya.cli.single.CliReporter;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.reporter.Problem;
 import org.aya.util.reporter.Reporter;
+import org.aya.util.reporter.ThrowingReporter;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.TestOnly;
 
@@ -17,6 +17,6 @@ import java.io.PrintStream;
 @TestOnly
 public record StreamReporter(@NotNull PrintStream stream) implements Reporter {
   @Override public void report(@NotNull Problem problem) {
-    stream.println(CliReporter.errorMessage(problem, DistillerOptions.informative(), false, false, 80));
+    stream.println(ThrowingReporter.errorMessage(problem, DistillerOptions.informative(), false, false, 80));
   }
 }
